lib/deploy: Add flag ostree-kargs-override to bootconfig
Add an ostree-kargs-override flag which is set to true if the
kargs of the current deployment were overridden. The flag is
written to the bootconfig, and persists in subsequent deployments.

This is in preparation for #1836 which needs a way to determine
whether the previous deployment's kargs were modified by the user,
so that user-modified kargs can be carried forward to the next
deployment.