lib/deploy: Add flag ostree-kargs-generated-from-config to bootconfig
Add an ostree-kargs-generated-from-config flag which is set to false
if the kargs of the current deployment were overridden. The flag is
written to the bootconfig, and persists in subsequent deployments.

If the flag was not present before, assume that an override happened,
and write the ostree-kargs-generated-from-config=false flag.

This is in preparation for adding default kargs support which will
involve generating kargs for the next BLS config from config files
in the filesystem, instead of carrying over kargs from the
previous deployment.